reworking tool dialog
