+++ requirements.txt
@@ -1,2 +1,2 @@
 PyQt5==5.15.4
-xappt==0.6.0
+#xappt==0.6.0
