+++ resources/ui/tool_interface.ui
@@ -10,7 +10,7 @@
     <x>0</x>
     <y>0</y>
     <width>676</width>
-    <height>380</height>
+    <height>462</height>
    </rect>
   </property>
   <property name="windowTitle">
