+++ xappt_qt/gui/widgets/__init__.py
@@ -2,4 +2,4 @@ from .file_edit import FileEdit
 from .check_list import CheckList
 from .error_label import ErrorLabel
 from .text_edit import TextEdit
-from .tool_page import ToolPage
+
