+++ xappt_qt/gui/widgets/tool_page/converters/__init__.py
@@ -1,0 +1,6 @@
+from xappt_qt.gui.widgets.tool_page.converters.base import ParameterWidgetBase
+from xappt_qt.gui.widgets.tool_page.converters.convert_bool import ParameterWidgetBool
+from xappt_qt.gui.widgets.tool_page.converters.convert_float import ParameterWidgetFloat
+from xappt_qt.gui.widgets.tool_page.converters.convert_int import ParameterWidgetInt
+from xappt_qt.gui.widgets.tool_page.converters.convert_list import ParameterWidgetList
+from xappt_qt.gui.widgets.tool_page.converters.convert_str import ParameterWidgetStr
